Added USERID to article schema, integrated recommendation functionality
Changes include:

Created a class called FormalArticle that
1) changed class syntax to the more verbose version
2) added a recommendation button next to the save/delete buttons, removing the need for a seperate page for recommendations.

Replaced all references to Article in SavedArticles.js w/ FormalArticle.

Updated Article schema so that it takes in one additional parameter, username. This is so that we can load articles specific to the user.

May revert this later and simply store saved article ids as an array in user object. Would be more efficient as database grows larger.  Just realized this now