fixed navbar bug
Problem: When the user logs in, the login page sets the jwt token. However, this change is not picked up by the nav component. The nav component decides whether to render the loggedin nav or the loggedout nav only once (when the component is loaded), and therefore did not pick up changes made in another component.

Solution:
Created a state in the main App.js and a method to update this state w/ the jwt token.
Passed this method to the login via the router. The login calls the method when the login process is complete.

Passed the App's state as a prop to the navbar. The nav's state is now syncronized w/ the app's state, which in turn is in sync w/ whether the jwt token is present or not.